# specs/001-meal-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OptiMeal API - Meal Management
  description: |
    API for the OptiMeal meal management feature (001-meal-management).
    Provides endpoints for pantry inventory, meal logging, meal planning, feedback collection, and nutritional insights.
  version: 1.0.0
  contact:
    name: OptiMeal Team
    email: api@optimeal.com
servers:
  - url: https://api.optimeal.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                full_name:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  token_type:
                    type: string
                    example: bearer
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials

  # ===========================
  # Food Items
  # ===========================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # ===========================
    # FoodItem
    # ===========================
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...